First step to enable Poolformer, make it possible to generate a full ViT including the PatchEmbedding part
