+++ xformers/factory/block_factory.py
@@ -5,63 +5,26 @@
 
 
 import logging
-from dataclasses import asdict, dataclass
-from enum import Enum
-from typing import Any, Dict, Optional, Tuple, Union
+from dataclasses import asdict
+from typing import Optional, Tuple, Union
 
 import torch
 import torch.nn as nn
 
 from xformers.components import (
     LayerNormStyle,
+    PatchEmbeddingConfig,
     PostNorm,
     PreNorm,
     Residual,
     build_multi_head_attention,
+    build_patch_embedding,
 )
-from xformers.components.feedforward import (
-    FEEDFORWARD_REGISTRY,
-    FeedforwardConfig,
-    build_feedforward,
-)
-from xformers.components.positional_embedding import (
-    POSITION_EMBEDDING_REGISTRY,
-    PositionEmbeddingConfig,
-    build_positional_embedding,
-)
+from xformers.components.feedforward import build_feedforward
+from xformers.components.positional_embedding import build_positional_embedding
 from xformers.components.residual import get_deepnorm_coefficients
 from xformers.components.simplicial_embedding import SimplicialEmbedding
-from xformers.utils import generate_matching_config
-
-
-class LayerPositionBitmask(int, Enum):
-    First = 0b01
-    Last = 0b10
-    Default = 0b11
-
-
-class LayerPosition:
-    """Bitmask to mark this layer as first, last, nothing or both"""
-
-    def __init__(self):
-        self.bitmask = LayerPositionBitmask.Default
-
-    def is_first(self):
-        return bool(self.bitmask & LayerPositionBitmask.First)
-
-    def is_last(self):
-        return bool(self.bitmask & LayerPositionBitmask.Last)
-
-    def mark_not_first(self):
-        self.bitmask &= ~LayerPositionBitmask.First
-
-    def mark_not_last(self):
-        self.bitmask &= ~LayerPositionBitmask.Last
-
-
-class BlockType(str, Enum):
-    Encoder = "encoder"
-    Decoder = "decoder"
+from xformers.factory.block_configs import xFormerDecoderConfig, xFormerEncoderConfig
 
 
 def _get_ln_factory(
@@ -116,180 +79,6 @@ def _get_ln_factory(
     return ln_factory
 
 
-@dataclass(init=False)  # handle constructors explicitly to force type changes
-class xFormerBlockConfig:
-    """
-    The configuration structure to define a Transformer block.
-    This base class is applicable to both encoder and decoder definitions.
-
-    This completely defines each of the blocks, for instance in terms of dimensions,
-    position encoding, pre or post layer norms or reversibility.
-    """
-
-    dim_model: int
-    feedforward_config: FeedforwardConfig
-    position_encoding_config: Optional[PositionEmbeddingConfig]
-    block_type: BlockType
-    layer_norm_style: LayerNormStyle
-    layer_position: LayerPosition
-    use_triton: bool
-    reversible: bool
-    num_layers: int
-
-    def __init__(
-        self,
-        dim_model: int,
-        feedforward_config: Dict[str, Any],
-        position_encoding_config: Optional[Dict[str, Any]],
-        block_type: BlockType,
-        layer_norm_style: LayerNormStyle = LayerNormStyle("post"),
-        reversible: bool = False,
-        num_layers: int = 1,
-        layer_position: Optional[LayerPosition] = None,
-    ):
-
-        self.dim_model = dim_model
-        self.block_type = block_type
-        self.layer_norm_style = layer_norm_style
-        self.reversible = reversible
-        self.num_layers = num_layers
-
-        # Fill in possible gaps in the config for subparts of the block
-        self.feedforward_config = generate_matching_config(
-            feedforward_config,
-            FEEDFORWARD_REGISTRY[feedforward_config["name"]].config,
-        )
-
-        self.position_encoding_config = (
-            generate_matching_config(
-                position_encoding_config,
-                POSITION_EMBEDDING_REGISTRY[position_encoding_config["name"]].config,
-            )
-            if position_encoding_config is not None
-            else None
-        )
-
-        # Default is that this layer is the only one, so both first and last
-        if layer_position:
-            self.layer_position = layer_position
-        else:
-            self.layer_position = LayerPosition()
-
-
-@dataclass(init=False)
-class xFormerEncoderConfig(xFormerBlockConfig):
-    """
-    The configuration structure for an encoder block
-    """
-
-    multi_head_config: Dict[str, Any]
-    use_triton: bool
-    simplicial_embeddings: Optional[Dict[str, Any]]
-
-    def __init__(
-        self,
-        dim_model: int,
-        feedforward_config: Dict[str, Any],
-        multi_head_config: Dict[str, Any],
-        position_encoding_config: Optional[Dict[str, Any]] = None,
-        layer_norm_style: str = "post",
-        use_triton: bool = True,
-        simplicial_embeddings: Optional[Dict[str, Any]] = None,
-        **kwargs,
-    ):
-        # Convenience, fill in duplicated field
-        try:
-            if "dim_model" not in multi_head_config.keys():
-                multi_head_config["dim_model"] = dim_model
-
-            if "dim_model" not in feedforward_config.keys():
-                feedforward_config["dim_model"] = dim_model
-
-            if (
-                position_encoding_config is not None
-                and "dim_model" not in position_encoding_config.keys()
-            ):
-                position_encoding_config["dim_model"] = dim_model
-
-        except AttributeError:
-            # A config instance was passed in, this is fine
-            pass
-        if "block_type" in kwargs:
-            assert kwargs["block_type"] == "encoder"
-        kwargs["block_type"] = BlockType("encoder")
-        super().__init__(
-            dim_model=dim_model,
-            feedforward_config=feedforward_config,
-            position_encoding_config=position_encoding_config,
-            layer_norm_style=LayerNormStyle(layer_norm_style),
-            **kwargs,
-        )
-
-        self.multi_head_config = multi_head_config
-        self.use_triton = use_triton
-        self.simplicial_embeddings = simplicial_embeddings
-
-
-@dataclass(init=False)
-class xFormerDecoderConfig(xFormerBlockConfig):
-    """
-    The configuration structure for a decoder block.
-
-    This specifically defines the masked and cross attention mechanisms,
-    on top of the settings defining all blocks.
-    """
-
-    multi_head_config_masked: Dict[str, Any]  # prior to encoder output
-    multi_head_config_cross: Dict[str, Any]  # cross attention, takes encoder output
-
-    def __init__(
-        self,
-        dim_model: int,
-        feedforward_config: Dict[str, Any],
-        multi_head_config_masked: Dict[str, Any],
-        multi_head_config_cross: Dict[str, Any],
-        position_encoding_config: Optional[Dict[str, Any]] = None,
-        layer_norm_style: str = "post",
-        use_triton: bool = True,
-        **kwargs,
-    ):
-
-        # Convenience, fill in duplicated field
-        try:
-            if "dim_model" not in multi_head_config_masked.keys():
-                multi_head_config_masked["dim_model"] = dim_model
-
-            if "dim_model" not in multi_head_config_cross.keys():
-                multi_head_config_cross["dim_model"] = dim_model
-
-            if "dim_model" not in feedforward_config.keys():
-                feedforward_config["dim_model"] = dim_model
-
-            if (
-                position_encoding_config is not None
-                and "dim_model" not in position_encoding_config.keys()
-            ):
-                position_encoding_config["dim_model"] = dim_model
-        except AttributeError:
-            # A config instance was passed in, this is fine
-            pass
-        if "block_type" in kwargs.keys():
-            assert kwargs["block_type"] == "decoder"
-        kwargs["block_type"] = BlockType("decoder")
-
-        super().__init__(
-            dim_model=dim_model,
-            feedforward_config=feedforward_config,
-            position_encoding_config=position_encoding_config,
-            layer_norm_style=LayerNormStyle(layer_norm_style),
-            **kwargs,
-        )
-
-        self.multi_head_config_masked = multi_head_config_masked
-        self.multi_head_config_cross = multi_head_config_cross
-        self.use_triton = use_triton
-
-
 class xFormerEncoderBlock(torch.nn.Module):
     r"""A vanilla Transformer Encoder block"""
 
@@ -314,11 +103,9 @@ class xFormerEncoderBlock(torch.nn.Module):
             mha_dim = config.multi_head_config["dim_model"]
 
             if pos_encoding_dim != mha_dim:
-
                 logging.warning(
                     f"The embedding dim and model dim do not match ({pos_encoding_dim} vs {mha_dim}), adding a projector layer."  # noqa
                 )
-
                 self.embedding_projector = nn.Linear(pos_encoding_dim, mha_dim)
         else:
             self.pose_encoding = None
@@ -362,6 +149,17 @@ class xFormerEncoderBlock(torch.nn.Module):
                 **config.simplicial_embeddings
             )
 
+        # Optional patch embedding
+        self.patch_emb: Optional[nn.Module] = None
+
+        if (
+            config.patch_embedding_config is not None
+            and config.layer_position.is_first()
+        ):
+            self.patch_emb = build_patch_embedding(
+                PatchEmbeddingConfig(**config.patch_embedding_config)
+            )
+
     @classmethod
     def from_config(cls, config: xFormerEncoderConfig):
         return cls(config)
@@ -388,6 +186,9 @@ class xFormerEncoderBlock(torch.nn.Module):
         att_mask: Optional[torch.Tensor] = None,
         input_mask: Optional[torch.Tensor] = None,
     ):
+        if self.patch_emb is not None:
+            x = self.patch_emb(x)
+
         if self.pose_encoding is not None:
             x = self.pose_encoding(x)
